main.swift -> Week1_Step2.swift 파일명 변경
