--- CodeStarterCamp_Week1/Week1_Step2_Dino.swift
@@ -1,8 +1,0 @@
-//
-//  Week1_Step2_Dino.swift
-//  CodeStarterCamp_Week1
-//
-//  Created by Jun on 2022/02/15.
-//
-
-import Foundation
